chore: redirect root page and remove unused components
