+++ .env.example
@@ -45,5 +45,5 @@ SGID_PRIVATE_KEY=
 
 # Other settings
 # ================
-# NEXT_PUBLIC_APP_NAME= # Uncomment to set application name
+NEXT_PUBLIC_APP_NAME="Isomer Studio (EXAMPLE)"
 # NEXT_PUBLIC_APP_URL= # Uncomment to set application URL
