+++ .env.test
@@ -3,5 +3,6 @@
 NODE_ENV=test
 DATABASE_URL=postgres://root:root@localhost:5432/test
 SESSION_SECRET=random_session_secret_that_is_at_least_32_characters
+NEXT_PUBLIC_APP_NAME="Isomer Studio (TEST)"
 NEXT_PUBLIC_S3_ASSETS_DOMAIN_NAME=user-content.example.com
 NEXT_PUBLIC_S3_ASSETS_BUCKET_NAME=assets-bucket
