--- apps/studio/src/schemas/thread.ts
@@ -1,9 +1,0 @@
-import { z } from "zod"
-
-import { addPostSchema } from "./post"
-
-export const addReplySchema = addPostSchema.extend({
-  postId: z.string(),
-})
-
-export type AddReplySchema = z.infer<typeof addReplySchema>
