+++ apps/studio/src/templates/layouts/AdminLayout.tsx
@@ -2,6 +2,7 @@ import { Flex } from "@chakra-ui/react"
 
 import { AppNavbar } from "~/components/AppNavbar"
 import { EnforceLoginStatePageWrapper } from "~/components/AuthWrappers"
+import { LayoutHead } from "~/components/LayoutHead"
 import { APP_GRID_COLUMN } from "~/constants/layouts"
 import { type GetLayout } from "~/lib/types"
 import { AppGrid } from "../AppGrid"
@@ -9,6 +10,8 @@ import { AppGrid } from "../AppGrid"
 export const AdminLayout: GetLayout = (page) => {
   return (
     <EnforceLoginStatePageWrapper>
+      <LayoutHead />
+
       <Flex minH="$100vh" flexDir="column" bg="base.canvas.alt" pos="relative">
         <AppNavbar />
         <AppGrid flex={1}>
