+++ apps/studio/src/templates/layouts/AdminSidebarOnlyLayout.tsx
@@ -14,6 +14,7 @@ import { z } from "zod"
 import type { CmsSidebarItem } from "~/components/CmsSidebar/CmsSidebarItems"
 import { EnforceLoginStatePageWrapper } from "~/components/AuthWrappers"
 import { CmsSidebar, CmsSidebarOnlyContainer } from "~/components/CmsSidebar"
+import { LayoutHead } from "~/components/LayoutHead"
 import { useMe } from "~/features/me/api"
 import { useQueryParse } from "~/hooks/useQueryParse"
 import { type GetLayout } from "~/lib/types"
@@ -21,6 +22,7 @@ import { type GetLayout } from "~/lib/types"
 export const AdminSidebarOnlyLayout: GetLayout = (page) => {
   return (
     <EnforceLoginStatePageWrapper>
+      <LayoutHead />
       <Flex minH="$100vh" flexDir="row" bg="base.canvas.alt" pos="relative">
         <CmsSidebarWrapper>{page}</CmsSidebarWrapper>
       </Flex>
