--- apps/studio/src/templates/layouts/ThreadLayout.tsx
@@ -1,3 +1,0 @@
-import { AdminLayout } from "./AdminLayout"
-
-export const ThreadLayout = AdminLayout
